Updating current project to the templates
