+++ src/shoppingTwo/shop/static/css/styles.css
@@ -1,5 +1,1 @@
-.sidebar-nav {
-    margin-top: 20px;
-    padding: 0;
-    list-style: none;
-}
+
