+++ src/shoppingTwo/shop/static/js/something.js
