+++ src/shoppingTwo/shop/templates/favorite_items.html
