+++ src/shoppingTwo/shop/urls.py
@@ -2,21 +2,33 @@ from django.urls import path
 from . import views
 
 urlpatterns = [
-    path("", views.homepage, name="homepage")
+    path("", views.homepage.as_view(), name="homepage")
 ]
 
 urlpatterns += [
-    path('cart1/', views.cartOne, name="cartOne"),
+    path('homepage/', views.homepage.as_view(), name="homepage"),
 ]
 
 urlpatterns += [
-    path('cart2/', views.cartTwo, name="cartTwo"),
+    path('store_items/', views.store_items, name="store_items"),
 ]
 
 urlpatterns += [
-    path('cart3/', views.cartThree, name="cartThree"),
+    path('profile/', views.profile, name="profile"),
 ]
 
 urlpatterns += [
-    path('map/', views.map, name="map"),
+    path('profile/favoriteCart', views.favorite_cart, name="favorite_cart"),
+]
+
+urlpatterns += [
+    path('profile/favoriteItems', views.favorite_items, name="favorite_items"),
+]
+
+urlpatterns += [
+    path('map/', views.map.as_view(), name="map"),
+]
+
+urlpatterns += [
+    path('signin/', views.signin, name="signin"),
 ]
