--- src/shoppingTwoTemplate/shop/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class ShopConfig(AppConfig):
-    name = 'shop'
