--- src/shoppingTwoTemplate/shop/static/css/styles.css
@@ -1,1 +1,0 @@
-
