--- src/shoppingTwoTemplate/shop/templates/base_generic.html
@@ -1,76 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    {% block title %}<title>Down the Aisle</title>{% endblock %}
-    <meta charset="utf-8">
-    {% load static %}
-    <link rel="icon" href="{% static "images/icoDraft.ico" %}">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/css/bootstrap.min.css"
-          integrity="sha384-MCw98/SFnGE8fJT3GXwEOngsV7Zt27NXFoaoApmYm81iuXoPkFOJwJ8ERdknLPMO"
-          crossorigin="anonymous">
-    <!-- Add additional CSS in static file -->
-</head>
-
-<body>
-<nav class="navbar navbar-expand-lg navbar-dark bg-dark">
-    <div class="container-fluid d-flex justify-content-between">
-        <a class="navbar-brand"><span class="text-light"></span></a>
-        <div class="collapse navbar-collapse" id="navbarNavAltMarkup">
-            <div class="navbar-nav">
-                <form class="form-inline">
-                    <a href="{% url 'homepage' %}">
-                        <svg xmlns="http://www.w3.org/2000/svg"
-                             width="0.69448in" height="0.652811in"
-                             viewBox="0 0 100 94">
-                            <path id="Selection #2"
-                                  fill="#D5D8CB" stroke="black" stroke-width="1"
-                                  d="M 22.00,12.00
-           C 22.00,12.00 22.00,18.00 22.00,18.00
-             22.14,25.43 24.04,25.95 31.00,26.00
-             31.00,26.00 82.00,26.00 82.00,26.00
-             82.00,26.00 67.69,53.00 67.69,53.00
-             64.84,58.26 64.23,61.51 58.00,62.00
-             61.08,69.25 53.38,69.67 55.00,62.00
-             55.00,62.00 27.00,62.00 27.00,62.00
-             28.49,69.07 22.51,69.07 24.00,62.00
-             18.43,61.07 18.10,59.38 18.00,54.00
-             18.00,54.00 18.00,35.00 18.00,35.00
-             18.00,35.00 19.00,16.00 19.00,16.00
-             13.86,16.00 7.98,17.64 6.00,12.00
-             6.00,12.00 22.00,12.00 22.00,12.00 Z"/>
-                        </svg>
-                    </a> <!-- URL goes in href! -->
-                    <a href="{% url 'signin' %}"
-                       class="btn btn-outline-success py-2 d-none d-md-inline-block"
-                       role="button"
-                       style="margin:10px;">Login</a> <!-- URL goes in href! -->
-                    <a href="{% url 'store_items' %}"
-                       class="btn btn-outline-success py-2 d-none d-md-inline-block"
-                       role="button"
-                       style="margin:10px;">Items</a> <!-- URL goes in href! -->
-                    <a href="{% url 'map' %}"
-                       class="btn btn-outline-success py-2 d-none d-md-inline-block"
-                       role="button"
-                       style="margin:10px;">Map</a> <!-- URL goes in href! -->
-                    <a href="{% url 'profile' %}"
-                       class="btn btn-outline-success py-2 d-none d-md-inline-block"
-                       role="button"
-                       style="margin:10px;">My Profile</a> <!-- URL goes in href! -->
-                </form>
-            </div>
-        </div>
-    </div>
-</nav>
-
-
-{% block content %}{% endblock %}
-<footer class="footer">
-    <div class="container-fluid">
-        <span class="text-muted">Leftover5&trade; 2018</span>
-    </div>
-</footer>
-</body>
-
-</html>
