--- src/shoppingTwoTemplate/shop/templates/homepage.html
@@ -1,43 +1,0 @@
-{% extends "base_generic.html" %}
-
-{% block content %}
-    {% load static %}
-    <body>
-    <div class="jumbotron">
-        <h1 class="display-4">Welcome to Down the Aisle!</h1>
-        <p class="lead">Get around your favorite stores faster than ever before!</p>
-        <hr class="my-5">
-        <p>Tell us what's on your grocery list and we'll show you the best way to get around</p>
-        <p class="lead">
-        </p>
-    </div>
-    <div class="container">
-        {% for store in store_list %}
-            <div class="col-lg-2 justify-center">
-                <button class="btn btn-success btn-block">{{ store.storename }}</button>
-            </div>
-            <div class="col-lg-10">
-                <h6>{{ store.storename }}
-                    <small> is a national supermarket chain that has your back when you're craving your favorite
-                        foods.
-                    </small>
-                </h6>
-            </div>
-            <hr class="my-5">
-        {% endfor %}
-        </div>
-    </div>
-
-    <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js"
-            integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo"
-            crossorigin="anonymous"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.3/umd/popper.min.js"
-            integrity="sha384-ZMP7rVo3mIykV+2+9J3UJ46jBk0WLaUAdn689aCwoqbBJiSnjAK/l8WvCWPIPm49"
-            crossorigin="anonymous"></script>
-    <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/js/bootstrap.min.js"
-            integrity="sha384-ChfqqxuZUCnJSK3+MXmPNIyE6ZbWh2IMqE241rYiqJxyMiZ6OW/JmZQ5stwEULTy"
-            crossorigin="anonymous"></script>
-
-    </body>
-
-{% endblock %}
